+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 dd mmm yyyy
 
 - New: error limit for apps.
+- BC break: the Group <-> Player relation for group applications is replaced by its own object
+  (this DELETES all existing group applications).
 - Layout fixes.
 - Added player ID to account name everywhere.